pyproject.toml
[project]
name = "stb-new"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.11",
    "asyncio>=3.4.3",
    "fastapi>=0.115.7",
    "google-genai>=0.6.0",
    "google-generativeai>=0.8.4",
    "httpx>=0.28.1",
    "pydantic-settings>=2.7.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]
